Address a tricky little bug where the container's view frame rectangle would get "silently" corrupted when calling scrollToVisible().
(Because view frames are cached, changing the frame in this manner would cause the now-corrupted value to be returned to callers, until they explicitly set the frame to something else.)

To see this bug in action pre-fix, you can do the following:

  -  Create a scroll view and make a reference to it called 'scroll'
  -  Have at least one view offscreen, calling it 'lastItemView'
  -  In your console, enter:  scroll.scrollToView(lastItemView)

This would work.  However, then use the mouse wheel to scroll up to the top of the view and enter:

scroll.scrollToView(lastItemView)

...again.  It won't work the first time (because the calculations rely on the value that was corrupted by the first invocation of scrollToView) but it will work if you invoke it again.

This led to one "lost" up/down keyboard press if autocomplete users used the mouse wheel at all.