After some debating, restructured the text selection support from a simple property that used observers to update its value to a (cacheable) dynamic property, with a separate mutator function, setSelection().  In the process, I also removed the "update selection" coalescing.
The downside to this is that it makes the API slightly more complicated -- users can no longer simply get and set 'selection'.  Rather, we're now more like fieldValue/setFieldValue.

But the performance benefit is worth it.  Previously, even if no one was interested in the selection, we'd update the value (which involves allocating a new object) for each keystroke, etc.  With this new structure, we only pay the penalty if someone is listening for the value.