Don't trust the hashing values for objects to imply that we're dealing with the same object, because objects created using beget() all currently have the same guid value.
Unfortunately, this means that most of the previous 550ms --> 150ms speedup was bogus, because many methods weren't actually getting the proper displayName value.  But the code complication is still worth it, since we're now at ~ 430ms on my machine; a more modest gain, but still enough to warrant the changes.