Implemented error storage/retrieval mechanism on the store.
Errors passed into the store via the error callbacks (dataSourceDidError()
and dataSourceDidErrorQuery()) are now written to internal arrays
(recordErrors and queryErrors, respectively).  These errors can then be
retrieved via calls to readError() and readQueryError().  Also modified the
errorObject computed properties on SC.Record and SC.RecordArray to invoke
these functions.  This way, when a record or record array transitions to the
error state, a simple get('errorObject') call will return the relevant
SC.Error object for inspection.

The readError() and readQueryError() functions on SC.NestedStore are simply
delegated to the parent store.