Removed default top-border on ToolbarView's CSS.
Leaves the view with only a default background color, which is usually something the developer overrides. I removed the border on account that I never used borders in my use of ToolbarView, and I was tired of redundant overriding via custom CSS. I hope this is something other developers find useful.