Adds SC.browser.chrome support.
Now, this some kind of a problem. In the current implemented way SC.browser.safari and SC.browser.chrome will have a value when browsing on chrome. This way we can make sur that everything still runs. But still, if you need to figure out if you're running on chrome, you can do it by now.