Observable: Make logged changes fully inspectable on the console
By passing a second argument to console.log, we allow users to
interactively inspect the new value on the console, which is
particularly useful for simple structures (such as frames and layouts)
which do not normally print in a useful fashion.

I don't know about the portability implications of this patch.