Correct a deficiency in the nested store's implementation of writeDataHash() where writeStatus() -- or any other method that involved calling writeDataHash() without a hash -- would mark the record as editable but not actually break the 'dataHashes' prototype chain for the storeKey.  (The code seems to have been copy/pasted from SC.Store#writeDataHash.)
This would lead to the situation where the storeKey is marked as 'editable' but all the data is still being inherited from the parent store.  Amongst other problems, if the parent store changed the data, the nested store's record wouldn't be notified because _notifyRecordPropertyChange() would see the non-'inherited' storeKeyEditState and skip the notification.