Added parameter to SC.$ hasClass for requiring all to match
In 1.4 hasClass would only return YES if all items had the class.
Now that we have switched to jQuery, we are using the jQuery version
which returns YES if any item has the given class. You can now pass
YES as the second parameter to hasClass to get the 1.4 behavior.