Add a handlebars framework and test SC.TemplateView working with it.
NOTE: The handlebars framework uses the built single-file (browser) version of Handlebars. However, the build process used by the Jison executable insists on adding CommonJS declarations. Since abbot's regular expression for detecting requires is extremely loose, it matches require("sys").*, and abbot emits warnings. As a result, we have manually removed the CommonJS exports. We will be submitting a patch tomorrow to Jison to expose the non-CommonJS compile option to the CLI tool, which will make it possible to automatically import handlebars. We will also be adding a task to SproutCore to automatically import jQuery and Handlebars.