If a superclass was used to load a record first (ex.  toOne: {MyApp.Person}), then if a subclass later appeared with the same ID (ex. store.find(MyApp.Male, id), it would error out because the subclass would generate a new store key and as it passed the storeKey up to its superclass, the superclass would already have a different storeKey and throw an exception.  Plus store.recordTypesByStoreKey would forever be stuck on the superclass Record Type.
Instead of generating the key at the bottom and passing it up, I generate the key at the top and recursively pass it back.  That way if a subclass claims the ID later, it can get the current key from its superclass and claim the Record Type for store.recordTypesByStoreKey.