Fixed <rdar://problem/9511032>.
The app does not have focus on first load. It appears that the
browser does not fire the focus event on first load. Therefore,
ensure the focus handler is called in setup. The downside is
that the app looks focus on first load even if the window is not
in focus.