Made MenuPane's autoresizing opt-in.
It used to be that MenuPane and MenuItemView incorporated autoresizing
functionality by reopening the classes. This was not ideal because it
forced all classes inheriting from either to include capability to auto
resize by default. Now, there are two new classes: AutoResizingMenuPane
and AutoResizingMenuItemView that provide the same functionality. This
way you can opt-in instead of opt-out.