Made SC.MenuPane and SC.PickerPane SC.ScrollView aware.
When popup is called on SC.MenuPane or SC.PickerPane, with an anchor
view, the pane is not repositioned if the anchor view's position is
changed by scrolling. This will result in the pane being in the wrong
position.

The solution to this problem is the following:
In popup, if an anchor view is provided, traverse it's parent view
hierarchy to check if it is a within a scroll view. If a scroll view is
found, then observe its horizontalScrollOffset and verticalScrollOffset.
When either of those change, reposition the pane. This way the pane
moves with it's anchor view.