SC.Store.commitRecords() does throw error when encounters empty record.
commitRecords() function was throwing NOT_FOUND_ERROR when encountered
record with EMPTY status. This was not meaningful behaviour: for example
when you do unloadRecords() with recordType parameter and then call
commitRecords() you'd end up with NOT_FOUND_ERROR. But EMPTY status
is does not imply that anything went wrong and should be silently ignored.