Fix SC.TextFieldView to insert a new line when the enter key is press on a text area
A fix were already there for the SC.InlineTextFieldView class so I copy
it into the TextFieldView class. I've updated the
SC.InlineTextFieldView class so that it now call the super class in the
case where it is a text area.

I've also removed some useless code in the InlineTextFieldView class:
- A fix in the insertNewline method which is not needed anymore
- The method fieldValueDidChange which only call the super class