Fixes SC.MenuPane pageDown and pageUp.
Previously, they estimated the number of items in a page, but because items can have custom heights, this could become a poor estimation.  Instead, we simply iterate away from the current item until we find that the next iteration's item would be beyond a page scroll and stop.  This approach fixes the failing unit tests, because the sample menu has separator items in it with height only 9.

This closes #815