Fix memory leak in SC.Object.extend.
When ever a class was extended, the new subclass would be stored in the
subclasses property on the parent class. This meant that every class ever
created always had a reference being maintained to it, and could never
be garbage collected.

Since it is very common to extend a view class when creating views in order
to set layout and other properties, you can quickly end up with a very
large number of throwaway subclasses that never get collected.

The only usage of this subclasses Set is in expandedRecordTypes in
SC.Record, so I moved it to that class. While this does not completely
eliminate the memory leaks, it does reduce them, since SC.Record subclasses
likely are not anonymously extended.