SC.ContainerView: Add an option to reuse instantiated views
While I was going to make this pull request, I saw that there were
already one for this: #657

I still publish it since the approach is different. The idea is to save
all the views instantiated by the container and simply reuse it when
they need to be show again. When reuseInstantiatedViews is set to true,
the views are never destroys except is the container is destroy. In
this case only the views instantiated by the container are destroys.

I've set the property reuseInstantiatedViews to YES by default. Let me
know if this is a problem.
I also can change the name reuseInstantiatedViews if it is not explicit
enough.