Adds automatic design mode class name addition.  If the pane implements designModes and the current design mode is 'large', the pane and its childViews will get a 'large' class name added.  This allows the developer to write CSS like this:
    .my-view {
      &.small { line-height: 24px; }
      &.large { line-height: 36px; }
    }

I found that almost every view in a design mode of the app needed small adjustments to their CSS for the current mode.  While it was possible to use classNameBindings on to the 'designMode' property, it ended up being used on almost every child view which meant a lot of repeated typing, plus classNameBindings would have resulted in an additional observer on each child view.  Since design mode changes are distinct events in the app, we don't need to bind them up and can do it much more efficiently with a direct approach.