Adds SC.appCache and SC.platform.supportsApplicationCache.
Working with the application cache is a confusing and time-consuming task.  This simple object abstracts out the most important properties for developers to use: hasNewVersion, isNewVersionValid, progress and isReadyForOffline all the while using the minimum number of event listeners which are properly cleaned up when unneeded.  It also has a property shouldPoll that will lazily look for updates in the background at a fixed interval.  This could be especially useful in a test environment, where you want to ensure that the testers aren't running the old version (which happens frequently when using the application cache).