Fixes memory leak in SC.Set.
The way that SC.Set removes objects is to "untrack" the internal index of the object by shrinking its length, but it never actually removed the object at the last index.  The only way that the object could be freed is if a new object is inserted at the same internal index, thus replacing it.  If the objects were removed in the reverse order that they were added, every object would still be in the set (until they were possibly overwritten).