Fixes a bug that left childView elements in the DOM when they were rendered as part of their parent's layer and later removed.
If childView layers are rendered when the parentView's  layer is created, the `layer` property on the childView will not be cached.  What occurs is that if the childView is then removed from the parent view without ever having its `layer` requested, when it comes time to destroy the DOM layer of the childView, it will try to find it with a `get('layer')`.  The bug was that it only returned a layer if the view has a parent view.  However, since the child was removed from the parent first and then destroyed, it no longer has a parent view and would not try to find its leftover DOM layer.

- Includes unit test.