Fix an issue where the transformed value was not sync with the "from" object if the transform function has changed it
Bug reported here: #859

This fix could lead to infinit loops with existing apps if the
transformed value is not meant to be sync back with the "from" object
and if the developer forgot to use oneWay instead of from to define the
binding. For this, I've add a developer warning saying that it should
use oneWay.

I've also add the syncTransformedValue parameter to the transform
method. For complex case, where we use isForward to transform the
property differently between the two binded objects, it can be usefull
to disable the synchronisation.

I've updated all the predefined transform functions to use oneWay
instead of from. In fact, I believe oneWay should have been use
initially. Because of this bug, there were no difference using from or
oneWay.