Completely refactors the transition implementation for SC.ContainerView in order to properly support multiple nowShowing changes with transition animations.
After struggling to make the container view handle multiple nowShowing changes while the previous items were still exiting, I had to step back and draw up the proper statechart for the container view to try and figure out why I couldn't get certain transitions to behave exactly as they should.  I determined that the container view's states are very basic, but what was missing was that each content view needed its own statechart in order to manage its own state.  For instance, a content view may be told to exit multiple times while it is in the process of exiting and the number of times it is told to exit may effect its transition.  Therefore, I added a simple statechart, SC.ContainerContentStatechart, that the container view can create for each new content and give control of the content's transitioning to.  This makes it possible to do very complex transitions, such as the PUSH transition that adjusts its exit animation accordingly to account for being pushed off before even being fully pushed on.