Changes the internal classNames of SC.Theme to be an Array rather than a CoreSet.  Having it as a Set makes no sense and the amount of property overloading is ridiculous.  We have to check for Objects, Arrays, Strings and Sets, which is a waste of code rather than just standardizing on an Array like classNames is on SC.View.  Also, this causes problems if developers use the theme's classNames and assume it will be an Array like with SC.View.
This is a potentially incompatible change, but it actually fixes current code that uses the theme's classNames with the updated SC.RenderContext API so it should potentially work with any current code.  Also, because this API is so under-used, hopefully it will not affect any implementations.  Though, if it does cause actual backwards incompatibility, it will have to be reverted for 1.10.