Remove _recalculateDivider flag from SC.SplitView.
This flag does not serve any purpose. Having a percentage value for the
default thickness is already handled and does not require renderLayout
to be executed again.

In addition, this was causing a bug when double-clicking on the divider
as the first action after the split view was rendered. The new thicknesses
would be set, then renderLayout would execute because of this flag,
resetting them to the back to the default.