Binding synchronisation fix
Fix an issue with binding where the transformed value was not sync back
if the transform function has changed it.

Bug reported here: #859

This fix can lead existing apps to an infinit loops if the transformed
value is not meant to be sync back with the "from" object and if the
developer forgot to use oneWay instead of from to define the
binding. For this, I've add a developer warning telling him use oneWay.

I've also add the syncTransformedValue parameter to the transform
method. For complex case, where we use isForward to transform the
property differently between the two binded objects, it can be usefull
to disable the synchronisation.