Fix inline text field position
Since this commit:
https://github.com/sproutcore/sproutcore/commit/4d199900fcd8a60f757c10a8
2de01b5f0576a907 (which were welcome since besides fixing the z-index,
it also fix a problem where if an inline text field was in a scroll
view, the field wasn't moving when scrolling), the calculation of
position of the editor was completely wrong. This can be see in the
showcase, with the SC.LabelView and the SC.ListView examples.
Based on my app and the showcase app, this commit fix the problem. I
haven't added units test because I believe this can't be tested.
Note that I've also dropped IE7 support for this since I'm not able to
test it and because even Twitter Bootstrap will drop IE7 support in
there next release.