Fix SC.GridView insertionPoint position
Don't if it was working for someone but in my case, the insertionPoint
was not appended.
While I was fixing it, I find out that there is no style for it and
that the style for sc-list-insertion-point is only set in the
legacy_theme.
I'd love to add theme for it but I don't know where is the best place
to do so: legacy_theme, ace or directly in the ressource directory by
creating a grid.css
Maybe it could also be great to move the theme for
sc-list-insertion-point.
I'm waiting for work feedback before going any further.