Throw an exception if we try to find an itemView with a negative contentIndex
This can happen when trying to get the itemView of an object that is
not in the collection.
If we don't throw an exception, an empty view with an id like sc123--1
is created which lead to an exception saying that an item view
(sc123--1) was found while the item view does not exist. See that issue
for more details:  https://github.com/sproutcore/sproutcore/issues/943