target/action support for the experimental SC.SelectView
If the action was defined on the SC.SelectView, the action wasn't call
when a menu item was selected.

I was looking to add test for this but they are all broken. I get this
strange error in the console: Uncaught ReferenceError: YES is not
defined
Any clue on how to fix this ?