Refactors the SC.PickerPane positioning code for SC.PICKER_POINTER and SC.MENU_POINTER.
This code was essentially unreadable and did not even really work.  There were lots of odd constants that affected the positioning and placement in strange ways.  With this refactor, the picker pane properly positions itself on the most appropriate side and will slide itself up/down or left/right if it can in order to fit.

- adds ability for pointer to automatically adjust its position as the pane shifts left/right or up/down regardless of its size
- fixes pane positioning to take into account the borderFrame
- adds private `windowPadding` property that ensures pickers are positioned a certain amount of distance from all edges of the screen
- deprecates the extraRightOffset property since the pointer positions itself
- fixes bug with render delegate that left the position class on the view after it changed