Fix text field selectionDirection unit tests.
PhantomJS apparently does not support selectionDirection on text fields.
This revealed a small bug when the browser does not support it.
Even though the direction would not be set on the input, the
SC.TextSelection would be cached with the selectionDirection.

I changed this to set selectionDirection on the SC.TextSelection to 'none'
if the browser does not support it. I also added an SC.platform test for
this feature. Lastly, I changed the unit test to pass if the browser does
not support selectionDirection.