Fixes fading for SC.OverlayScrollerView in browsers that can’t transition opacity of child nodes with 3D translations on them. Since the thumb portion of the scroller is translated and since it is not an SC.View, we can’t call .animate on it directly and calling .animate on the actual parent node’s SC.View fails to work too.
Also, changes the fade out implementation slightly to only call fadeIn/fadeOut once on the ScrollView per change, which speeds things up a lot and saves the ScrollView implmentation from having to track state. Also ensures that the scrollviews don’t only fade out on append, but also fade back in on re-append or re-show.

Also, renames fadeDelay to fadeOutDelay to be more explicit on what the delay is (fade in vs. fade out) and removes fadeInDuration/fadeOutDuration, because the fade animation is strictly handled by the scroller.