Fixes an issue with SC.offset relative to the offset parent
jQuery fail to compute the correct offset when using $.position.
This is because, in SC, centered elements are positioned in percentage
like this: top: 50%; margin-top: -10px. And jQuery is subtracting the
margin from the offset which lead to a wrong value.
This commit fixe this and remove some hacky code in
positionOverTargetView which was related to this behaviour. This hack
were causing trouble when an SC.InlineTextFieldView were displayed from
an element positioned without percentage (like in SC.TableView).