Trigger observers whenever a record array changes
When a record array is empty because the backing query is still
in flight, setting the record array as content of an array controller
will cause the status to become READY_CLEAN despite the query being
in flight, as it doesn't find any records in the store. When the
store indicates it has fetched the query, the status will stay
READY_CLEAN and observers late to the chase (such as stateObservers)
will not be notified when the data comes in. Calling the
this.notifyPropertyDidChange() for the status causes these
observers to be fired.